Rename fulldecay to fromdecay and FullDecay to MultiDecayChain
