+++ tests/fromdecay/__init__.py
@@ -1,4 +1,4 @@
 import pytest
 
 # This makes it so that assert errors are more helpful for e.g., the check_norm helper function
-pytest.register_assert_rewrite("fulldecay.test_fulldecay")
+pytest.register_assert_rewrite("fromdecay.test_fulldecay")
