+++ tests/fromdecay/example_decays.dec
