+++ tests/fromdecay/test_fromdecay.py
@@ -1,13 +1,13 @@
 from numpy.testing import assert_almost_equal
 
-from phasespace.fulldecay import FullDecay
-from phasespace.fulldecay.mass_functions import _DEFAULT_CONVERTER
+from phasespace.fromdecay import MultiDecayChain
+from phasespace.fromdecay.mass_functions import _DEFAULT_CONVERTER
 
-from .example_decay_chains import *  # TODO remove * since it is bad practice?
+from . import example_decay_chains
 
 
-def check_norm(full_decay: FullDecay, **kwargs) -> list[tuple]:
-    """Helper function that checks whether the normalize_weights argument works for FullDecay.generate.
+def check_norm(full_decay: MultiDecayChain, **kwargs) -> list[tuple]:
+    """Helper function that checks whether the normalize_weights argument works for MultiDecayChain.generate.
     Args:
         full_decay: full_decay.generate will be called.
         kwargs: Additional parameters passed to generate.
@@ -33,7 +33,9 @@ def check_norm(full_decay: FullDecay, **kwargs) -> list[tuple]:
 
 def test_single_chain():
     """Test converting a decaylanguage dict with only one possible decay."""
-    container = FullDecay.from_dict(dplus_single, tolerance=1e-10)
+    container = MultiDecayChain.from_dict(
+        example_decay_chains.dplus_single, tolerance=1e-10
+    )
     output_decay = container.gen_particles
     assert len(output_decay) == 1
     prob, gen = output_decay[0]
@@ -56,7 +58,9 @@ def test_single_chain():
 
 def test_branching_children():
     """Test converting a decaylanguage dict where the mother particle can decay in many ways."""
-    container = FullDecay.from_dict(pi0_4branches, tolerance=1e-10)
+    container = MultiDecayChain.from_dict(
+        example_decay_chains.pi0_4branches, tolerance=1e-10
+    )
     output_decays = container.gen_particles
     assert len(output_decays) == 4
     assert_almost_equal(sum(d[0] for d in output_decays), 1)
@@ -66,7 +70,7 @@ def test_branching_children():
 
 def test_branching_grandchilden():
     """Test converting a decaylanguage dict where children to the mother particle can decay in many ways."""
-    container = FullDecay.from_dict(dplus_4grandbranches)
+    container = MultiDecayChain.from_dict(example_decay_chains.dplus_4grandbranches)
     output_decays = container.gen_particles
     assert len(output_decays) == 4
     assert_almost_equal(sum(d[0] for d in output_decays), 1)
@@ -77,9 +81,9 @@ def test_branching_grandchilden():
 
 def test_mass_converter():
     """Test that the mass_converter parameter works as intended."""
-    dplus_4grandbranches_massfunc = dplus_4grandbranches.copy()
+    dplus_4grandbranches_massfunc = example_decay_chains.dplus_4grandbranches.copy()
     dplus_4grandbranches_massfunc["D+"][0]["fs"][-1]["pi0"][-1]["zfit"] = "rel-BW"
-    container = FullDecay.from_dict(
+    container = MultiDecayChain.from_dict(
         dplus_4grandbranches_massfunc,
         tolerance=1e-10,
         mass_converter={"rel-BW": _DEFAULT_CONVERTER["relbw"]},
@@ -98,7 +102,7 @@ def test_mass_converter():
 
 
 def test_big_decay():
-    container = FullDecay.from_dict(dstarplus_big_decay)
+    container = MultiDecayChain.from_dict(example_decay_chains.dstarplus_big_decay)
     output_decays = container.gen_particles
     assert_almost_equal(sum(d[0] for d in output_decays), 1)
     check_norm(container, n_events=1)
